ApprovedRevs only recognises per-page rights, not group rights
When testing to see if a user has the approverevisions right, the only
check is of $title->userCan('approverevisions'). This check, though
has hooks that allow other code to alter the value returned on a
per-page basis.

Instead, $wgUser->isAllowed('approverevisions') should be used since
it is a general check on the approverevisions right.

Change-Id: I77e3c4bbe0b73b6d3992eaf41dbc6eec2ebbc82f