Allow approving file revisions
File upload versions can now be marked approved. This is different from
approving revisions of file *pages*. Revisions to the wikitext of file
*pages* cannot be marked approved, to avoid confusion with approval of
the file upload revisions.

Approving file upload revisions is done on the file page itself (e.g. on
File:Test.jpg directly, versus clicking on "View history" on that page)
in the "File history" section. Once a file revision is approved, that
version of the file will show up wherever the file is used. For images,
the approved revision of the image will show on other pages. For
non-images, any direct links to the file will lead to the approved
revision.

The file namespace is now approvable by default. To remove it, add the
following to LocalSettings.php:

$egApprovedRevsEnabledNamespaces[NS_FILE] = false;

Change-Id: I6785dc3a3d8023ad9a7c917fb91cb8f7250d34b7