Make userCanApprove() able to cache multiple user/page combos
Previously userCanApprove() set self::$mUserCanApprove to true or false
to cache the result for just the logged in user and the current page in
context. However, to use userCanApprove() for any arbitrary user/page
combination this will not work. Instead, this change makes
self::$mUserCanApprove into an array with keys indicating the user/page
combination and values indicating approvability.

Change-Id: I17b86f3da085fd51acfb96fdfb4623b6b864a2d0