Add five new variables/parser functions for displaying approval info
The five new magic words are: {{APPROVALYEAR}}, {{APPROVALMONTH}}, {{APPROVALDAY}}, {{APPROVALTIMESTAMP}} and {{APPROVALUSER}} - they are analogous to the REVISION...  magic words from core MediaWiki. Each one can be called by itself or with an argument (a page name), which means that each one is both a variable and a parser function.

Change-Id: I91c63bcd1b875eedaebe9f5cd98938a49af38e41